pharmaceutical-innovation-clusters/requirements.txt
# Data Collection
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.4

# Analysis
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# LLM APIs (for potential future enhancements)
openai==1.54.3

# Utilities
python-dotenv==1.0.0
tqdm==4.67.1

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
